gateway/requirements.txt
flask
flask-cors
influxdb-client
python-dotenv